+++ src/main/java/org/swarg/mc/traprescue/RescueLogger.java
@@ -11,4 +11,8 @@ public class RescueLogger {
     public static void logInfo(String msg) {
         LOG.info("[TrapRescue] {}", msg);
     }
+
+    public static void logWarn(String msg) {
+        LOG.warn("[TrapRescue] {}", msg);
+    }
 }